requirements.txt
python-dotenv==1.0.0
Requests==2.28.2
